+++ protocols/encoding/encoding.go
@@ -12,21 +12,28 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package encodings
+package encoding
 
-import "go.opentelemetry.io/collector/consumer/pdata"
+import "fmt"
 
-// MetricsEncoder serializes pdata into bytes.
-type MetricsEncoder interface {
-	EncodeMetrics(md pdata.Metrics) ([]byte, error)
+// Type is the encoding format that a model is serialized to.
+type Type string
+
+const (
+	Protobuf Type = "protobuf"
+	JSON     Type = "json"
+	Thrift   Type = "thrift"
+)
+
+func (e Type) String() string {
+	return string(e)
 }
 
-// TracesEncoder serializes pdata into bytes.
-type TracesEncoder interface {
-	EncodeTraces(td pdata.Traces) ([]byte, error)
+// ErrUnavailableEncoding is returned when the requested encoding is not present.
+type ErrUnavailableEncoding struct {
+	Encoding Type
 }
 
-// LogsEncoder serializes pdata into bytes.
-type LogsEncoder interface {
-	EncodeLogs(ld pdata.Logs) ([]byte, error)
+func (e *ErrUnavailableEncoding) Error() string {
+	return fmt.Sprintf("unsupported encoding %q", e.Encoding)
 }
